@@ -15,30 +15,37 @@
     "write-translations": "docusaurus write-translations",
     "write-heading-ids": "docusaurus write-heading-ids",
     "lint": "eslint --ext .js,.jsx,.ts,.tsx . --fix",
-    "lint:content": "eslint --no-eslintrc -c .content.eslintrc.js content --fix"
+    "lint:content": "eslint --no-eslintrc -c .content.eslintrc.js content --fix",
+    "diagram2code:generate": "docusaurus diagram2code:generate"
   },
   "dependencies": {
     "@babel/preset-react": "^7.18.6",
     "@cloudinary/react": "^1.11.2",
     "@cloudinary/url-gen": "^1.9.2",
-    "@docusaurus/core": "2.4.3",
-    "@docusaurus/preset-classic": "2.4.3",
-    "@docusaurus/remark-plugin-npm2yarn": "2.4.3",
-    "@medusajs/icons": "^1.0.0",
+    "@docusaurus/core": "latest",
+    "@docusaurus/preset-classic": "latest",
+    "@docusaurus/remark-plugin-npm2yarn": "latest",
+    "@docusaurus/theme-mermaid": "latest",
+    "@mdx-js/mdx": "3.0.0",
+    "@mdx-js/react": "3",
+    "@medusajs/icons": "1.2.1",
     "@svgr/webpack": "6.2.1",
+    "@swc/core": "^1.3.102",
     "autoprefixer": "^10.4.14",
     "clsx": "^1.1.1",
     "docs-ui": "*",
-    "docusaurus-plugin-image-zoom": "^0.1.1",
+    "docusaurus-plugin-image-zoom": "^1.0.1",
     "dotenv": "^16.0.3",
     "eslint": "^8.49.0",
     "eslint-config-docs": "*",
     "file-loader": "^6.2.0",
     "lodash": "^4.17.21",
+    "mdast-util-mdx": "^3.0.0",
+    "meilisearch-docsearch": "^0.6.0",
     "postcss": "^8.4.21",
-    "prism-react-renderer": "^1.3.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "prism-react-renderer": "^2.3.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-tooltip": "5.7.4",
     "react-transition-group": "^4.4.5",
     "tailwind": "*",
@@ -60,9 +67,16 @@
   },
   "devDependencies": {
     "@babel/plugin-proposal-decorators": "^7.22.15",
-    "@docusaurus/module-type-aliases": "2.4.3",
-    "@tsconfig/docusaurus": "^1.0.7",
+    "@docusaurus/module-type-aliases": "latest",
+    "@docusaurus/tsconfig": "latest",
+    "@docusaurus/types": "latest",
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
     "@types/react-transition-group": "^4.4.6",
-    "typescript": "^5.0.2"
+    "swc-loader": "^0.2.3",
+    "typescript": "~5.2.2"
+  },
+  "engines": {
+    "node": ">=18.0"
   }
 }
