@@ -33,40 +33,45 @@
     "dev": "yarn build:js:cjs && yarn build:js:esm"
   },
   "devDependencies": {
-    "@types/react": "^17.0.1",
-    "@types/react-dom": "^17.0.1",
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
     "@types/react-google-recaptcha": "^2.1.6",
     "clsx": "^2.0.0",
     "cpy-cli": "^5.0.0",
     "eslint-config-docs": "*",
-    "next": "latest",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "next": "14.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "rimraf": "^5.0.1",
     "tailwind": "*",
     "tailwindcss": "^3.3.3",
     "tsc-alias": "^1.8.7",
     "tsup": "^5.10.1",
+    "types": "*",
     "typescript": "^5.1.6"
   },
   "peerDependencies": {
-    "@types/react": ">=17",
-    "@types/react-dom": ">=17",
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
     "next": "*",
-    "react": ">=17",
-    "react-dom": ">=17"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "dependencies": {
-    "@medusajs/icons": "latest",
-    "@medusajs/ui": "latest",
+    "@medusajs/icons": "^1.2.0",
+    "@medusajs/ui": "^2.4.1",
     "@octokit/request": "^8.1.1",
     "@react-hook/resize-observer": "^1.2.6",
     "@segment/analytics-next": "^1.55.0",
+    "@uidotdev/usehooks": "^2.4.1",
     "algoliasearch": "^4.20.0",
-    "prism-react-renderer": "^2.0.6",
+    "mermaid": "^10.9.0",
+    "npm-to-yarn": "^2.1.0",
+    "prism-react-renderer": "2.3.1",
     "react-google-recaptcha": "^3.1.0",
     "react-instantsearch": "^7.0.3",
     "react-markdown": "^8.0.7",
+    "react-medium-image-zoom": "^5.1.10",
     "react-tooltip": "^5.21.3",
     "react-transition-group": "^4.4.5",
     "react-uuid": "^2.0.0"
