@@ -1,18 +1,19 @@
 import React from "react"
-import type { Props } from "@theme/DocItem/Layout"
+
+interface Props {
+  data: string[]
+}
 
 export default function UnorderedCodeList({ data }: Props): JSX.Element {
   return (
-        <ul>
-        {
-          data.map((token, i) => {
-            return(
-              <li>
-                <code>{token}</code>
-              </li>
-            )
-          })
-        }
+    <ul>
+      {data.map((token: string, index: number) => {
+        return (
+          <li key={index}>
+            <code>{token}</code>
+          </li>
+        )
+      })}
     </ul>
   )
 }
