@@ -7,8 +7,8 @@ import isInternalUrl from "@docusaurus/isInternalUrl"
 import IconExternalLink from "@theme/Icon/ExternalLink"
 import type { Props } from "@theme/DocSidebarItem/Link"
 import { ModifiedPropSidebarItemLink } from "@medusajs/docs"
-import DocSidebarItemIcon from "../../../components/DocSidebarItemIcon"
 import { Badge } from "docs-ui"
+import DocSidebarItemIcon from "../../../components/DocSidebarItemIcon"
 
 type ModifiedProps = Props & {
   item: ModifiedPropSidebarItemLink
@@ -24,6 +24,7 @@ export default function DocSidebarItemLink({
   const { href, label, className, autoAddBaseUrl, customProps } = item
   const isActive = isActiveSidebarItem(item, activePath)
   const isInternalLink = isInternalUrl(href)
+
   return (
     <li
       className={clsx(
@@ -67,6 +68,12 @@ export default function DocSidebarItemLink({
         )}
         {label}
         {!isInternalLink && <IconExternalLink />}
+        {customProps?.sidebar_badge && (
+          <Badge
+            {...customProps.sidebar_badge}
+            className={`sidebar-soon-badge`}
+          />
+        )}
       </Link>
       {customProps?.sidebar_is_soon && (
         <Badge variant="purple" className={`sidebar-soon-badge`}>
