@@ -20,7 +20,6 @@ import {
   CashSolid,
   Channels,
   ChannelsSolid,
-  ChatBubbleLeftRight,
   CheckCircleSolid,
   CheckMini,
   ChevronDoubleLeftMiniSolid,
@@ -42,12 +41,10 @@ import {
   CubeSolid,
   CurrencyDollar,
   CurrencyDollarSolid,
-  CursorArrowRays,
   Discord,
   DocumentText,
   DocumentTextSolid,
   EllipseMiniSolid,
-  Envelope,
   ExclamationCircle,
   ExclamationCircleSolid,
   FlyingBox,
@@ -55,7 +52,6 @@ import {
   FolderOpen,
   Gatsby,
   GiftSolid,
-  Github,
   GlobeEurope,
   GlobeEuropeSolid,
   InformationCircleSolid,
@@ -64,16 +60,13 @@ import {
   KeySolid,
   LightBulb,
   LightBulbSolid,
-  Link,
   Linkedin,
   MagnifyingGlass,
   Map,
   Moon,
   Newspaper,
   PencilSquareSolid,
-  PlayMiniSolid,
   Puzzle,
-  QuestionMarkCircle,
   ReactJsEx,
   ReceiptPercent,
   RocketLaunch,
@@ -83,7 +76,7 @@ import {
   ServerStackSolid,
   ShoppingCart,
   ShoppingCartSolid,
-  Sidebar,
+  SidebarLeft,
   Sparkles,
   SparklesSolid,
   SquareTwoStackSolid,
@@ -100,19 +93,22 @@ import {
   Twitter,
   User,
   UsersSolid,
-  Users,
   XCircleSolid,
   XMark,
   XMarkMini,
   PhotoSolid,
-  Heart,
-  Medusa,
   ListBullet,
   Swatch,
-  Trash,
+  Medusa,
+  Heart,
+  QuestionMark,
+  ChatBubbleLeftRight,
 } from "@medusajs/icons"
 import IconPuzzleSolid from "./PuzzleSolid"
 import IconNextjs from "./Nextjs"
+import IconFlagMini from "./FlagMini"
+import IconBeaker from "./Beaker"
+import IconGitHub from "./GitHub"
 import IconNestjs from "./Nestjs"
 import IconLaravel from "./Laravel"
 import IconGingonic from "./Gingonic"
@@ -142,6 +138,7 @@ export default {
   "arrow-down-tray": ArrowDownTray,
   "back-arrow": ArrowUturnLeft,
   "bars-three": BarsThree,
+  beaker: IconBeaker,
   bell: BellAlert,
   "bell-alert-solid": BellAlertSolid,
   bolt: Bolt,
@@ -156,7 +153,6 @@ export default {
   "cash-solid": CashSolid,
   "channels-solid": ChannelsSolid,
   channels: Channels,
-  "chat-bubble-left-right": ChatBubbleLeftRight,
   "check-circle-solid": CheckCircleSolid,
   "check-mini": CheckMini,
   "chevron-double-left-mini-solid": ChevronDoubleLeftMiniSolid,
@@ -180,13 +176,11 @@ export default {
   "cube-solid": CubeSolid,
   "currency-dollar": CurrencyDollar,
   "currency-dollar-solid": CurrencyDollarSolid,
-  "cursor-arrow-rays": CursorArrowRays,
   "dark-mode": Moon,
   discord: Discord,
   "document-text": DocumentText,
   "document-text-solid": DocumentTextSolid,
   "ellipse-mini-solid": EllipseMiniSolid,
-  "envelope": Envelope,
   "exclamation-circle-solid": ExclamationCircleSolid,
   "external-link": ArrowUpRightOnBox,
   "flying-box": FlyingBox,
@@ -194,7 +188,8 @@ export default {
   "folder-open": FolderOpen,
   gatsby: Gatsby,
   "gift-solid": GiftSolid,
-  github: Github,
+  "flag-mini": IconFlagMini,
+  github: IconGitHub,
   "globe-europe": GlobeEurope,
   "globe-europe-solid": GlobeEuropeSolid,
   "information-circle-solid": InformationCircleSolid,
@@ -204,7 +199,6 @@ export default {
   "light-bulb": LightBulb,
   "light-bulb-solid": LightBulbSolid,
   "light-mode": Sun,
-  "link": Link,
   linkedin: Linkedin,
   "magnifying-glass": MagnifyingGlass,
   map: Map,
@@ -212,9 +206,7 @@ export default {
   nextjs: IconNextjs,
   "pencil-square-solid": PencilSquareSolid,
   "photo-solid": PhotoSolid,
-  "play-mini-solid": PlayMiniSolid,
   puzzle: Puzzle,
-  "question-mark": QuestionMarkCircle,
   // TODO change once available in Icons package
   "puzzle-solid": IconPuzzleSolid,
   react: ReactJsEx,
@@ -227,7 +219,7 @@ export default {
   "server-stack-solid": ServerStackSolid,
   "shopping-cart": ShoppingCart,
   "shopping-cart-solid": ShoppingCartSolid,
-  sidebar: Sidebar,
+  sidebar: SidebarLeft,
   sparkles: Sparkles,
   "sparkles-solid": SparklesSolid,
   "squares-plus": SquaresPlus,
@@ -237,16 +229,16 @@ export default {
   stripe: Stripe,
   "swatch-solid": SwatchSolid,
   "tag-solid": TagSolid,
-  "trash": Trash,
   tools: Tools,
   "tools-solid": ToolsSolid,
   twitter: Twitter,
   user: User,
-  "users": Users,
   "users-solid": UsersSolid,
   "x-circle-solid": XCircleSolid,
   "x-mark": XMark,
   "x-mark-mini": XMarkMini,
+  "question-mark": QuestionMark,
+  "chat-bubble-left-right": ChatBubbleLeftRight,
   heart: Heart,
   nestjs: IconNestjs,
   laravel: IconLaravel,
