@@ -1,23 +1,26 @@
 import React, { type ComponentProps, type ReactElement } from "react"
+import Details from "@theme/Details"
 import type { Props } from "@theme/MDXComponents/Details"
-import Details, { DetailsProps } from "../Details"
-import { type DetailsSummaryProps } from "docs-ui"
+import { DetailsSummary } from "docs-ui"
 
-type MDXDetailsProps = Props & DetailsProps
-
-export default function MDXDetails(props: MDXDetailsProps): JSX.Element {
+export default function MDXDetails(props: Omit<Props, "key">): JSX.Element {
   const items = React.Children.toArray(props.children)
   // Split summary item from the rest to pass it as a separate prop to the
   // Details theme component
   const summary = items.find(
-    (item): item is ReactElement<ComponentProps<"summary">> =>
-      React.isValidElement(item) &&
-      (item.props as { mdxType: string } | null)?.mdxType === "summary"
-  ) as React.ReactElement<DetailsSummaryProps>
+    (
+      item: ReactElement<ComponentProps<"summary">>
+    ): item is ReactElement<ComponentProps<"summary">> => {
+      return (
+        React.isValidElement(item) &&
+        (item.type === "summary" || item.type === DetailsSummary)
+      )
+    }
+  )
   const children = <>{items.filter((item) => item !== summary)}</>
 
   return (
-    <Details summary={summary} {...props}>
+    <Details {...props} summary={summary}>
       {children}
     </Details>
   )
