@@ -1,17 +1,28 @@
 // Import the original mapper
 import MDXComponents from "@theme-original/MDXComponents"
 import CloudinaryImage from "@site/src/components/CloudinaryImage"
-import MDXDetails from "./Details"
 import MDXA from "./A"
-import { Kbd, DetailsSummary, InlineCode } from "docs-ui"
+import { Kbd, DetailsSummary, Note } from "docs-ui"
+import H1 from "./H1"
+import MDXCode from "./Code"
+import MDXDetails from "./Details"
+import DocCard from "@theme/DocCard"
+import DocCardList from "@theme/DocCardList"
 
-export default {
+const components = {
   // Re-use the default mapping
   ...MDXComponents,
-  inlineCode: InlineCode,
+  code: MDXCode,
   img: CloudinaryImage,
   details: MDXDetails,
-  summary: DetailsSummary,
+  Details: MDXDetails,
+  Summary: DetailsSummary,
   a: MDXA,
   kbd: Kbd,
+  h1: H1,
+  DocCard,
+  DocCardList,
+  Note,
 }
+
+export default components
