@@ -2,9 +2,9 @@ import React from "react"
 import TOCItems from "@theme-original/TOCItems"
 import type TOCItemsType from "@theme/TOCItems"
 import type { WrapperProps } from "@docusaurus/types"
-import StructuredDataHowTo from "@site/src/components/StructuredData/HowTo"
 import { useDoc } from "@docusaurus/theme-common/internal"
 import { DocContextValue } from "@medusajs/docs"
+import StructuredDataHowTo from "../../components/StructuredData/HowTo"
 
 type Props = WrapperProps<typeof TOCItemsType>
 
