@@ -5,8 +5,7 @@ import { IconProps } from "@medusajs/icons/dist/types"
 
 export type BorderedIconProps = {
   icon?: string
-  IconComponent?: ((props: IconProps) => React.JSX.Element) | null
-  // IconComponent?: React.FC<IconProps> | null
+  IconComponent?: React.FC<IconProps> | null
   wrapperClassName?: string
   iconWrapperClassName?: string
   iconClassName?: string
