@@ -1,34 +1,40 @@
 import React from "react"
-import { CodeBlock, InlineCode } from "@/components"
+import {
+  CodeBlock,
+  CodeBlockMetaFields,
+  InlineCode,
+  MermaidDiagram,
+} from "@/components"
+import { Npm2YarnCode } from "../Npm2YarnCode"
 
 export type CodeMdxProps = {
   className?: string
   children?: React.ReactNode
-}
+} & CodeBlockMetaFields
 
 // due to how mdx handles code blocks
 // it is required that a code block specify a language
 // to be considered a block. Otherwise, it will be
 // considered as inline code
-export const CodeMdx = ({ className, children }: CodeMdxProps) => {
+export const CodeMdx = ({ className, children, ...rest }: CodeMdxProps) => {
   if (!children) {
     return <></>
   }
 
   const match = /language-(\w+)/.exec(className || "")
 
+  const codeContent = Array.isArray(children)
+    ? (children[0] as string)
+    : (children as string)
+
   if (match) {
-    return (
-      <CodeBlock
-        source={
-          Array.isArray(children)
-            ? (children[0] as string)
-            : (children as string)
-        }
-        lang={match[1]}
-      />
-    )
+    if (rest.npm2yarn) {
+      return <Npm2YarnCode npmCode={codeContent} {...rest} />
+    } else if (match[1] === "mermaid") {
+      return <MermaidDiagram diagramContent={codeContent} />
+    }
+    return <CodeBlock source={codeContent} lang={match[1]} {...rest} />
   }
 
-  return <InlineCode>{children}</InlineCode>
+  return <InlineCode>{codeContent}</InlineCode>
 }
