@@ -3,13 +3,15 @@ import clsx from "clsx"
 import { PlusMini } from "@medusajs/icons"
 
 export type DetailsSummaryProps = {
-  title: React.ReactNode
-  subtitle?: string
+  title?: React.ReactNode
+  subtitle?: React.ReactNode
   badge?: React.ReactNode
   expandable?: boolean
   open?: boolean
   className?: string
   titleClassName?: string
+  hideExpandableIcon?: boolean
+  summaryRef?: React.LegacyRef<HTMLDivElement>
 } & Omit<React.HTMLAttributes<HTMLElement>, "title">
 
 export const DetailsSummary = ({
@@ -21,6 +23,8 @@ export const DetailsSummary = ({
   open = false,
   className,
   titleClassName,
+  hideExpandableIcon = false,
+  summaryRef,
   ...rest
 }: DetailsSummaryProps) => {
   return (
@@ -28,10 +32,13 @@ export const DetailsSummary = ({
       className={clsx(
         "py-docs_0.75 flex items-center justify-between",
         expandable && "cursor-pointer",
-        !expandable && "border-medusa-border-base border-y",
+        !expandable &&
+          "border-medusa-border-base border-y border-solid border-x-0",
+        (expandable || badge) && "gap-0.5",
         "no-marker",
         className
       )}
+      ref={summaryRef}
       {...rest}
     >
       <span className="gap-docs_0.25 flex flex-col">
@@ -44,7 +51,7 @@ export const DetailsSummary = ({
           {title || children}
         </span>
         {subtitle && (
-          <span className="text-compact-medium text-medusa-fg-subtle">
+          <span className="text-compact-medium text-medusa-fg-subtle mt-0.5">
             {subtitle}
           </span>
         )}
@@ -52,7 +59,7 @@ export const DetailsSummary = ({
       {(badge || expandable) && (
         <span className="flex gap-docs_0.5">
           {badge}
-          {expandable && (
+          {expandable && !hideExpandableIcon && (
             <PlusMini
               className={clsx("transition-transform", open && "rotate-45")}
             />
