@@ -1,15 +1,22 @@
 import React from "react"
+import NextLink from "next/link"
+import type { LinkProps as NextLinkProps } from "next/link"
 import clsx from "clsx"
 
 export type LinkProps = {
   href?: string
   children?: React.ReactNode
   className?: string
-} & React.AllHTMLAttributes<HTMLAnchorElement>
+  target?: string
+  rel?: string
+} & Partial<NextLinkProps>
 
 export const Link = ({ href, children, className, ...rest }: LinkProps) => {
+  if (href?.replace(/#.*$/, "").endsWith("page.mdx")) {
+    href = href.replace("/page.mdx", "")
+  }
   return (
-    <a
+    <NextLink
       href={href || ""}
       {...rest}
       className={clsx(
@@ -18,6 +25,6 @@ export const Link = ({ href, children, className, ...rest }: LinkProps) => {
       )}
     >
       {children}
-    </a>
+    </NextLink>
   )
 }
