@@ -7,7 +7,7 @@ import { ModalHeader } from "./Header"
 import { ModalFooter } from "./Footer"
 import { useKeyboardShortcut } from "@/hooks/use-keyboard-shortcut"
 import { ButtonProps } from "@/components"
-import { Ref } from "@/types"
+import { Ref } from "types"
 
 export type ModalProps = {
   className?: string
@@ -20,7 +20,7 @@ export type ModalProps = {
   footerContent?: React.ReactNode
   passedRef?: Ref<HTMLDialogElement>
   headerClassName?: string
-} & Omit<React.ComponentProps<"dialog">, "ref" | "title">
+} & Omit<React.ComponentProps<"dialog">, "ref" | "title" | "key">
 
 export const Modal = ({
   className,
