@@ -36,8 +36,7 @@ export const NotificationItem = ({
   return (
     <div
       className={clsx(
-        "md:max-w-[320px] md:w-[320px] w-full bg-medusa-bg-base rounded-docs_DEFAULT",
-        "shadow-flyout dark:shadow-flyout-dark max-h-[calc(100vh-90px)]",
+        "md:max-w-[320px] md:w-[320px] w-full",
         "fixed md:right-docs_1 left-0 md:m-docs_1",
         placement === "bottom" && "md:bottom-docs_1 bottom-0",
         placement === "top" && "md:top-docs_1 top-0",
