@@ -3,7 +3,7 @@
 import React, { useCallback, useEffect, useRef } from "react"
 import clsx from "clsx"
 import { OptionType } from "@/hooks"
-import { Ref } from "@/types"
+import { Ref } from "types"
 import { CheckMini, EllipseMiniSolid } from "@medusajs/icons"
 
 export type SelectDropdownProps = {
