@@ -1,11 +1,10 @@
 "use client"
 
-import React, { useEffect, useState } from "react"
+import React, { useId } from "react"
 import { Tooltip as ReactTooltip } from "react-tooltip"
 import type { ITooltip } from "react-tooltip"
 import clsx from "clsx"
 import "react-tooltip/dist/react-tooltip.css"
-import uuid from "react-uuid"
 
 export type TooltipProps = {
   text?: string
@@ -24,13 +23,7 @@ export const Tooltip = ({
   className,
   ...tooltipProps
 }: TooltipProps) => {
-  const [elementId, setElementId] = useState<string>("")
-
-  useEffect(() => {
-    if (!elementId) {
-      setElementId(uuid())
-    }
-  }, [elementId])
+  const elementId = useId()
 
   return (
     <>
