@@ -3,22 +3,36 @@ export * from "./Bordered"
 export * from "./BorderedIcon"
 export * from "./Button"
 export * from "./Card"
+export * from "./CardList"
+export * from "./ChildDocs"
 export * from "./CodeBlock"
 export * from "./CodeMdx"
 export * from "./CodeTabs"
+export * from "./CodeTabs/Item"
 export * from "./CopyButton"
 export * from "./Details"
 export * from "./Details/Summary"
+export * from "./DetailsList"
+export * from "./EditButton"
+export * from "./ExpandableNotice"
+export * from "./FeatureFlagNotice"
+export * from "./HooksLoader"
 export * from "./Icons"
+export * from "./InlineIcon"
+export * from "./InlineThemeImage"
 export * from "./InlineCode"
 export * from "./Input/Text"
 export * from "./Kbd"
 export * from "./Label"
+export * from "./LegacyCodeTabs"
+export * from "./LegacyLink"
 export * from "./Link"
 export * from "./Loading"
 export * from "./Loading/Dots"
 export * from "./Loading/Spinner"
 export * from "./MarkdownContent"
+export * from "./MDXComponents"
+export * from "./MermaidDiagram"
 export * from "./Modal"
 export * from "./Modal/Header"
 export * from "./Modal/Footer"
@@ -29,13 +43,27 @@ export * from "./Navbar/Link"
 export * from "./Navbar/Logo"
 export * from "./Navbar/MobileMenu"
 export * from "./Navbar/MobileMenu/Button"
-export { Link as NextLink, type LinkProps as NextLinkProps } from "./NextLink"
+export * from "./Note"
 export * from "./Notification"
 export * from "./Notification/Item"
 export * from "./Notification/Item/Layout/Default"
-export * from "./Rating"
+export * from "./Pagination"
+export * from "./Search"
+export * from "./Search/EmptyQueryBoundary"
+export * from "./Search/Hits"
+export * from "./Search/Hits/GroupName"
+export * from "./Search/ModalOpener"
+export * from "./Search/NoResults"
+export * from "./Search/Suggestions"
+export * from "./Search/Suggestions/Item"
 export * from "./Select"
 export * from "./Sidebar"
 export * from "./Sidebar/Item"
+export * from "./Table"
+export * from "./Tabs"
 export * from "./TextArea"
+export * from "./ThemeImage"
+export * from "./Toggle"
 export * from "./Tooltip"
+export * from "./TypeList"
+export * from "./ZoomImg"
