@@ -1,2 +1,8 @@
-export const GITHUB_ISSUES_PREFIX = `https://github.com/medusajs/medusa/issues/new?assignees=&labels=type%3A+docs&template=docs.yml`
-export const GITHUB_UI_ISSUES_PREFIX = `https://github.com/medusajs/ui/issues/new?labels=documentation`
+import { NavbarLinkProps } from "@/components"
+import { SidebarItemType } from "types"
+
+export const GITHUB_ISSUES_PREFIX = `https://github.com/zeropsio/docs/issues/new`
+
+export const navbarItems: NavbarLinkProps[] = []
+
+export const mobileSidebarItems: SidebarItemType[] = []
