@@ -2,6 +2,6 @@ export * from "./constants"
 
 export * from "./components"
 export * from "./hooks"
+export * from "./layouts"
 export * from "./providers"
-export * from "./types"
 export * from "./utils"
