@@ -1,8 +1,9 @@
-export * from "./Analytics"
 export * from "./ColorMode"
 export * from "./Mobile"
 export * from "./Modal"
 export * from "./Navbar"
 export * from "./Notification"
 export * from "./PageLoading"
+export * from "./Pagination"
+export * from "./Search"
 export * from "./Sidebar"
